Created verse-utils.ts. Added translation term for reference format string.
